+++ biochar/__init__.py
@@ -21,8 +21,9 @@ from .biochar_generator import (
 )
 from .heteroatom_assignment import CompositionResult, CompositionInfo
 from .surface_builder import SurfaceBuilder, SurfaceConfig, SheetResult
+from .temperature_model import TemperatureModel, properties, VALID_FEEDSTOCKS
 
-__version__ = "0.1.4"
+__version__ = "0.1.5"
 __all__ = [
     "BiocharGenerator",
     "GeneratorConfig",
@@ -36,4 +37,7 @@ __all__ = [
     "SurfaceBuilder",
     "SurfaceConfig",
     "SheetResult",
+    "TemperatureModel",
+    "properties",
+    "VALID_FEEDSTOCKS",
 ]